LayoutTests/fast/css/ancestor-of-hovered-element-removed-expected.txt

Verify the hovered state is updated correctly when an ancestor of the hovered element is detached from the document's tree
On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

    Initial state
PASS elementsWithHoverStyle() is ["html", "body", "prime-ancestor"]
PASS elementsMatchingHoverSelector() is ["html", "body", "prime-ancestor"]
Moving over #target
PASS elementsWithHoverStyle() is ["html", "body", "prime-ancestor", "group", "element-to-remove", "target"]
PASS elementsMatchingHoverSelector() is ["html", "body", "prime-ancestor", "group", "element-to-remove", "target"]
Removing the renderer of #element-to-remove
PASS elementsMatchingHoverSelector(elementToRemove) is []
PASS elementsWithHoverStyle() is ["html", "body", "prime-ancestor", "interceptor"]
PASS elementsMatchingHoverSelector() is ["html", "body", "prime-ancestor", "interceptor"]
PASS successfullyParsed is true

TEST COMPLETE
